+++ packages/server/package.json
@@ -24,7 +24,9 @@
   "dependencies": {
     "@types/express": "4.17.2",
     "@types/http-proxy-middleware": "0.19.3",
+    "@types/multer": "^1.3.10",
     "@types/node": "13.1.6",
+    "@types/path-to-regexp": "^1.7.0",
     "@types/signale": "^1.2.1",
     "@types/sockjs": "0.3.31",
     "body-parser": "^1.19.0",
