+++ packages/server/src/mock/normalizeConfig.ts
@@ -1,0 +1,77 @@
+import assert from 'assert';
+import { RequestHandler } from 'express';
+import bodyParser from 'body-parser';
+import multer from 'multer';
+import { pathToRegexp } from 'path-to-regexp';
+
+const VALID_METHODS = ['get', 'post', 'put', 'patch', 'delete'];
+const BODY_PARSED_METHODS = ['post', 'put', 'patch', 'delete'];
+
+function parseKey(key: string) {
+  let method = 'get';
+  let path = key;
+  if (/\s+/.test(key)) {
+    const splited = key.split(/\s+/);
+    method = splited[0].toLowerCase();
+    path = splited[1]; // eslint-disable-line
+  }
+  assert(
+    VALID_METHODS.includes(method),
+    `Invalid method ${method} for path ${path}, please check your mock files.`,
+  );
+  return {
+    method,
+    path,
+  };
+}
+
+function createHandler(method: any, path: any, handler: any): RequestHandler {
+  return function(req, res, next) {
+    if (BODY_PARSED_METHODS.includes(method)) {
+      bodyParser.json({ limit: '5mb', strict: false })(req, res, () => {
+        bodyParser.urlencoded({ limit: '5mb', extended: true })(
+          req,
+          res,
+          () => {
+            sendData();
+          },
+        );
+      });
+    } else {
+      sendData();
+    }
+
+    function sendData() {
+      if (typeof handler === 'function') {
+        multer().any()(req, res, () => {
+          handler(req, res, next);
+        });
+      } else {
+        res.json(handler);
+      }
+    }
+  };
+}
+
+export default function normalizeConfig(config: any) {
+  return Object.keys(config).reduce((memo: any, key) => {
+    const handler = config[key];
+    const type = typeof handler;
+    assert(
+      type === 'function' || type === 'object',
+      `mock value of ${key} should be function or object, but got ${type}`,
+    );
+    const { method, path } = parseKey(key);
+    const keys: any[] = [];
+    // @ts-ignore
+    const re = pathToRegexp(path, keys);
+    memo.push({
+      method,
+      path,
+      re,
+      keys,
+      handler: createHandler(method, path, handler),
+    });
+    return memo;
+  }, []);
+}
